Apply specific order when adding token attributes with the identical key
If two token placeholders overlap, it was possible to get inconsistent
attributes, e.g. half the tokens with one value and half with the other.
Thus if two named entities overlapped, e.g. "Pakistan" and "Pakistan
International Airlines", we could get "place" for Pakistan but "company"
for International Airlines. This is now corrected: for a given key and
overlapping placeholders, only one of the placeholders will be applied.
Order is: smaller startIndex gets precedence, if equal greater endIndex
gets precedence.