Separate configuration from resources by removing language packs
The use of language packs confused configuration: they were supposed to
come between the reference config and the user-defined config. Instead,
added the possibility in the configuration file to refer to resources
inside a zip (or other) archive using Apache Commons VFS. This means the
language packs are replaced by a zip and a config file referring to the zip.

This allowed us also to get rid of LanguageImplementations (replaced by
config + TalismaneSession).

Finally, the idea of setting a "base directory" has been removed. Either
the configuration files use absolute paths (recommended), or they are
interpreted relative to the current working directory.