Transform TokenFilters into Annotators
Specifically, the TokenFilters now add two types of annotations: one
for TokenPlaceholders, which indicates where "hard-coded" tokens need to
be added to the sentence, with a possible replacement, and TokenAttributes
which will add arbitrary name/value attributes to any token covered by the annotation
span (but only one attribute of a given name per token).

Sentence has become an AnnotatedText to which these annotations are added.

The annotations are now automatically applied by the TokenSequence,
specifically when it is initially constructed (for TokenPlaceholders)
and whenever a token is added (for TokenAttribute). We are now guaranteed
that any token added will be assigned the correct attributes.

A positive side-effect is that PretokenisedSequence is now simplified, but
must receive a fully constructed sentence when constructed, hence
some organisational changes in the various regex-based annotated corpus
readers, to build a sentence before we create the PretokenisedSequence.