Create global set of annotators referenced everywhere
The annotators are initialized in TalismaneSession. We no longer read
annotators from the models - instead it is the responsibility of the
person distributing the model to distribute the annotators which the model
depends on to function correctly. This vastly simplifies the management
of annotators, since before we could have the same (or different) annotators
in each model, and when analysing with four models it was necessary to
decide which model's annotators to use and how to handle model annotators
vs annotators supplied in the configuration.

Moreover, annotators are configured at top-level, and are then applied
as appropriate by the various modules, depending on the command (analyse,
train, evaluate, compare). Generally, we attempt to apply the annotators
as early as possible after the object to be annotated has been created,
e.g. a sentence, a token sequence.