Change TextMarkerFilter to annotate a text block
Basically, the existing mechanism used a filter to add text markers, then
used the text markers to generate the processed text, then ran sentence
detection on the processed text, and then build sentences.

Changing this to: run filters on the raw text block, run sentence
detection on the raw text block, generate a sentence builder, and build
the sentences. This is an intermediate step, before skipping the
sentence builder completely.

In this new scenario, the sentence detector adds annotations just like the
filters.