Base Talismane processing on a rolling text block
Blocks of text get rolled in one at a time from right to left.
There are always 4 blocks.
The raw text filters are always applied to block 3, with block 4 as
context in case a regex crosses the boundary between blocks 3 and 4.
Whenever rolling, new new block 2, for which filters have already been
applied, is processed, giving us the processed text, ready for analysis.

We can then detect sentences on block 2, with blocks 1 and 3 as context,
in case features need context to the left or right of the block boundary.

The sentence detector now adds sentence boundary annotations to the
text being annotated.