Fix Talismane client/server functionality
This was completely missing from the TalismaneMain class.

Also, the end-of-input marks were causing errors, since annotations were
gettting processed twice.
In order to ensure a particular annotation didn't get processed twice,
it turned out to be simplest to add so-called "processsing marks"  which
allow us to indicate for each annotation whether it's beginning and end
have already been processed. These can get added but never removed. They
are transferred when the annotations are "rolled" to a new position.

Also, we preload modules in Talismane class constructor, to ensure they
get loaded before the first analysis is requested.