Make output processors configurable for all modules
Previously, configuration only allowed us to use pre-defined output
processors. We now allow any output processor, as long as it implements
the interface required by the module (e.g. ParseConfigurationProcessor for
the parser), and provides one of three constructors. This makes it much
simpler to extend Talimsne's processing capabilities, simply by adding
a class to the classpath which meets these constraints, and changing the
configuration file to point to this class.