Start sentences at first non space character
Typically, a punctuation mark is often followed by a space (e.g. a
period marking the end of a sentence in English). The sentence following
this punctuation starts immediately after, but we can consider that the
leading spaces are not relevant to the sentence boundary and that they
can be skipped.