Make evaluation observers fully configurable for all modules
Also made the Tokeniser, PosTagger and Parser themselves configurable.

This makes it much simpler to extend Talismane - one simply needs to
implement the required interfaces and create a constuctor with the
correct arguments.

Also, split the parser, posTagger and tokeniser packages into a evaluate
and output sub-package for readability, and to avoid clouding the core
code with ancillary functionality.