Read config from talismane.core.[sessionId] rather than talismane.core
This makes it possible to easily create multiple Talismane sessions in
a single app. If we want one Talismane in English and one in French, which
activate based on language detection, we simply create a configuration
based at "talismane.core.en" and another at "talismane.core.fr", and
open a TalismaneSession with session ids "en" and "fr" respectively.

It is simplest for the user to inherit his config from
talismane.core.generic.