Delete default logging configuration
When Talismane is used as a library, this default logging configuration with DEBUG
level for all joliciel classes pollutes the application logs. Moreover, it is not the role of a library
to define the default logging configuration with some log file and an arbitrary rolling
strategy, let the client application do that.